Add drag and drop and color choosing to rocplot.
The colors of individual curves are now shown in the right panel to give clearer association between the right panel and curves in the main plot.

It is now possible to drag and drop curves in the right panel to change the order of the curves in the plot.  The older up and down arrows have been retired.

A new menu item allows the user to change the color of any curve.