Improvements to decomposition
* More separation of the decomposition, and allow use as a VcfIterator
* Robustness in the face of invalid alleles
* Refinements to when decomposition is employed
* Allow decomposition without the reference SDF

The reference SDF is only required when decomposition produces a pure
indel at the very left of the original record, in order to provide the
reference base to the left for anchoring. In such situations an
alternative valid anchoring is to use the reference base to the right
(which is always available in the original record given that decomposition
has occurred).