vcfeval: enable precision/recall for sub-type ROC data files.
The header now reports the total baseline and total called variants as
the sub-type-specific totals. This means that we can calculate
more useful precision/recall/f-measure for the sub-type files.

The caveat is that since ROC score values are attached to calls,
categorization of true/false positives are necessarily determined using
the call representation. However, since sensitivity is defined as TP
/ (TP + FN), but FN sub-type classification is inherently tied to
baseline representation, the sensitivity metric can be biased if there
is some a consistent representational bias between baseline and
calls. These biases can be minimized by employing decomposition.