Apply rescaling to subcategory ROCs based on end-point correspondance.
Subcategory precision/recall curves could give strange results in cases
where there are different representational conventions between baseline
and called variants. Now we apply a scaling factor to the subcategory
baseline TP's such that the curve endpoint sensitivity corresponds to the
raw baseline per-category sensitivity.

Note that if you're plotting a subcategory which can't be meaningfully
evaluated for the baseline call set (e.g. XRX) the results may not be
what you expect, so there is an experimental flag to disable the
scaling.