rocplot: implement precision/sensitivity interpolation.
For sparsely populated precision-sensitivity graphs, it's incorrect to
connect adjacent points with a straight line, as this does not
correspond to the ROC convex hull and over-estimates the
accuracy. Instead, we can plot interpolated values, following
Davis (2006).

Note that to be consistent with interpolated points, the process of
labelling points on the curve has changed. Previously the label shown at
a point corresponded to the threshold that should be applied
exclusively. Now the threshold is inclusive. For densely populated
graphs this makes little visual difference.

Currently, interpolation is only enabled via experimental flags.