vcffilter: Add the ability for JavaScript to set ID, QUAL and FILTERs
From JavaScript, reading the FILTER object yields a list of the filters that
have been set. Any changes to this list are not reflected back into the
VCF record. Instead there are two ways that the VCF FILTER field may be
written to:

Direct assignment, of either a string containing semicolon separated
filters to set, or of a list of filters, e.g.:

    FILTER = 'BAZ;BANG';
    FILTER = ['BAZ', 'BANG'];

Adding a single filter to existing filters can be accomplished by
the above methods. E.g.: Adding a value to the existing list and setting
the result:

    var f = FILTER;
    f.push('BOING');
    FILTER = f;

Since this is a little unwieldy, there is a new function called add()
that can also be used (and may be chained):

    FILTER.add('BOING');
    FILTER.add(['BOING', 'DOING');
    FILTER.add('BOING').add('DOING');

Also added the ability to write to QUAL, e.g. QUAL = 50;

Also added the ability to write to ID as string or list, e.g.:

    ID = 'a;b';
    ID = ['a', 'b'];

Support setting ID