Fix reading of VCF records when there are multiple regions
This commit is a putative fix for a problem of records inadvertently
being skipped when reading VCF records with
SamMultiRestrictingIterator.  This problem definitely can occur in the
presence of placed unmapped records, but potentially also in other
situations.  It occurs because mPreviousAlignmentStart was updated to
the position of the first record outside of the current region, but
then in the next record, ALL records at that position would be stepped
over as having been processed, when in fact they had not yet been
processed.  This fix makes mPreviousAlignmentStart one less than it
would otherwise be.