vcfmerge: prevent crash in case of record containing duplicated alleles when using --preserve-formats
It's not actually clear what the semantics are for a record where one of
the ALTs is the same as the REF or if there are two copies of an
ALT neither are disallowed by the spec. vcfmerge now takes the swiss
approach and just passes these through without doing allele
normalization rather than crashing.