samplesim: New flag --allow-missing-af
When a record is encountered without AF annotations, the previous
version of samplesim would treat all alleles as equally likely during
simulation (e.g. equivalent to AF=0.5 for a biallelic variant). This
doesn't play well with denovosim in more complex simulations so the new
behaviour is to treat these as AF=0. The old behaviour is available via
the new flag.