vcfeval: Support "partial spanning deletion" notation.
This is where additional bases may be added to the spanning deletion
symbol in the case of overlapping indels. It is currently assumed that
any additional bases should only ever be the same as the reference (and
these are actually ignored for matching purposes).

Thanks to Dan Cooke for sending an example from Octopus of this.